pictures_api/requirements.txt
SQLAlchemy==2.0.23
Flask==2.3.2
imagekitio==3.0.1
requests==2.31.0
waitress==2.1.2
PyMySQL==1.0.3
cryptography==41.0.1
urllib3==1.26.16
Pillow==9.0.0
mysql-connector-python==8.2.0